requirements.txt
python-docx>=1.0.0
docxcompose>=1.4.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pytest>=8.0.0
